Abstract the concept of the consensus algorithm
This pr creates a new trait with the basic methods of a
consensus algorithm; sync with peer and run consensus.

It also changes the default implementation to explicitly be
swirlds.

Lachesis will implement this,
