+++ lachesis-rs/src/lib.rs
@@ -16,8 +16,10 @@ mod node;
 mod peer;
 mod printable_hash;
 mod round;
+mod swirlds;
 
 pub use hashgraph::{BTreeHashgraph, Hashgraph, HashgraphWire};
 pub use event::{Event, EventHash};
 pub use node::Node;
+pub use swirlds::Swirlds;
 pub use peer::{Peer, PeerId};
